+++ src/deck-editor.js
@@ -1,0 +1,21 @@
+import React, { Component } from "react";
+ 
+class DeckEditor extends Component {
+  render() {
+    return (
+      <div>
+        <h2>Deck Editor</h2>
+        <p>Choose a deck to edit:</p>
+
+        <li>
+            Insert Deck Name Here
+        </li>
+        <li>
+            Insert Deck Name Here
+        </li>
+      </div>
+    );
+  }
+}
+ 
+export default DeckEditor;
