+++ src/home.js
@@ -1,0 +1,21 @@
+import React, { Component } from "react";
+ 
+class Home extends Component {
+  render() {
+    return (
+      <div>
+        <h2>Home Page</h2>
+        <p>Choose a deck to play:</p>
+
+        <li>
+            Insert Deck Name Here
+        </li>
+        <li>
+            Insert Deck Name Here
+        </li>
+      </div>
+    );
+  }
+}
+ 
+export default Home;
