+++ src/main.js
@@ -1,0 +1,30 @@
+import React, { Component } from "react";
+import {
+    Route,
+    NavLink,
+    HashRouter
+} from "react-router-dom";
+import Home from "./home";
+import DeckEditor from "./deck-editor";
+
+class Main extends Component {
+    render() {
+        return (
+            <HashRouter>
+                <div>
+                    <h1>Simple SPA</h1>
+                    <ul className="header">
+                        <li><NavLink to="/">Home</NavLink></li>
+                        <li><NavLink to="/deck-editor">DeckEditor</NavLink></li>
+                    </ul>
+                    <div className="content">
+                        <Route exact path="/" component={Home} />
+                        <Route path="/deck-editor" component={DeckEditor} />
+                    </div>
+                </div>
+            </HashRouter>
+        );
+    }
+}
+
+export default Main;
